requirements.txt
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Login>=0.6,<1.0
Flask-WTF>=1.2,<2.0
email-validator>=2.0,<3.0
pg8000>=1.31,<2.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0
Flask-Limiter>=3.5,<4.0
Flask-Migrate>=4.0,<5.0
Flask-Mail>=0.10,<1.0
Flask-Cors>=4.0,<5.0
APScheduler>=3.10,<4.0
cryptography>=44.0,<45.0
PyJWT>=2.8,<3.0
openpyxl>=3.1,<4.0
